Version 3.2.1 (June 11, 2016) - added option to change database password and host - fixed partsreport subtotal grouping - updated character set list for postgresql - moved templates into directories for easier maintenance - added General Ledger import - fixed adding a new language - added Service Company COA for Germany - added missing to email for reminders - fixed foreign exchange for orders with prepayments - added maximum number of tries for one time TAN login - added units for timecard in generate sales order function - changed character encoding to select option - added payment info to print on order - added totals for income sheet and balance sheet when headings are not included - changed error reporting when creating a new dataset - added xelatex option to generate PDF files - added client_encoding to preferences

+++ SL/AM.pm
@@ -1059,9 +1059,9 @@ sub save_language {
   my $dbh = $form->dbconnect($myconfig);
 
   $form->{code} =~ s/ //g;
-  foreach my $item (qw(code description)) {
-    $form->{$item} =~ s/-(-)+/-/g;
-    $form->{$item} =~ s/ ( )+/-/g;
+  for (qw(code description)) {
+    $form->{$_} =~ s/-(-)+/-/g;
+    $form->{$_} =~ s/ ( )+/-/g;
   }
   
   # if there is an id
@@ -2516,7 +2516,9 @@ sub get_exchangerates {
 
   $form->{currencies} = $form->get_currencies($myconfig, $dbh);
 
-  ($form->{transdatefrom}, $form->{transdateto}) = $form->from_to($form->{year}, $form->{month}, $form->{interval}) if $form->{year} && $form->{month};
+  unless ($form->{transdatefrom} || $form->{transdateto}) {
+    ($form->{transdatefrom}, $form->{transdateto}) = $form->from_to($form->{year}, $form->{month}, $form->{interval}) if $form->{year} && $form->{month};
+  }
   
   $where .= " AND transdate >= '$form->{transdatefrom}'" if $form->{transdatefrom};
   $where .= " AND transdate <= '$form->{transdateto}'" if $form->{transdateto};   
@@ -3152,7 +3154,7 @@ sub restore {
   my @sql = <FH>;
   close(FH);
   
-  $file = "$myconfig->{dbdriver}-custom_tables.sql";
+  $file = "sql/$myconfig->{dbdriver}-custom_tables.sql";
   if (open(FH, "$file")) {
     push @sql, <FH>;
     close(FH);
@@ -3165,7 +3167,7 @@ sub restore {
 
   for (@sql) {
     
-    if (/references (\w+)/i) {
+    if (/references /i) {
       $references{$el} = 1;
     }
     
@@ -3188,10 +3190,13 @@ sub restore {
     return 0;
   }
 
+  $dbh->{PrintError} = 0;
+
   for ($dbh->tables) {
     if ($myconfig->{dbdriver} =~ /Pg/) {
       if (!/(pg_catalog|information_schema)/) {
-	$tables{$_} = 1;
+        $_ =~ s/public\.//;
+        $tables{$_} = 1;
       }
     } else {
       $tables{$_} = 1;
@@ -3203,11 +3208,13 @@ sub restore {
     $query = qq|DROP TABLE $_;|;
     $dbh->do($query);
   }
-  
+
   # drop tables and sequences
   for (keys %tables) {
-    $query = qq|DROP TABLE $_;|;
-    $dbh->do($query);
+    unless ($references{$_}) {
+      $query = qq|DROP TABLE $_;|;
+      $dbh->do($query);
+    }
   }
 
   # drop sequences
@@ -3221,24 +3228,31 @@ sub restore {
   close(FH);
   
   for (@sql) {
-    next if /--/;
+    next if /^--/;
 
     $query .= $_;
     
     if (/create function/i) {
       while (!/-- end function/i) {
-	$_ = shift @sql;
-	$query .= $_;
+        $_ = shift @sql;
+        $query .= $_;
       }
     }
     
     if (/;\s*$/) {
       if ($query =~ /VALUES/) {
-	next if $query !~ /\);$/;
+        next if $query !~ /\);$/;
       }
-      $query =~ s/;\s*$//;
+      $query =~ s/;(\s*)$//;
 
-      $dbh->do($query) or $form->info($DBI::errstr);
+      if ($query =~ /^DROP /) {
+        (undef, undef, $el) = split / /, $query;
+        if ($tables{$el} || $references{$el}) {
+          $query = "";
+          next;
+        }
+      }
+      $dbh->do($query) if $query;
       $query = "";
     }
   }
